Added some new methods to the classes
Added some new methods and restructured the code slightly for ease of use moving forwards.

Changes:

HazardManager
Changed so that a vector 2 is now set for each hazard to get their optimal range. Added method GetOptimalRange() to get these. Streamlined the Finished hazard method, removed it's switch statement for if a hazard is fixed and now calls a new method in the score script that takes in a string to set the subsequent boolean. Moved the HazardProgress() method to the hazard mechanics script.

HazardMechanics
Added a RunHazard() method which takes in a string and fetches the score values for that particular hazard as well as a float for progress on the hazard slider and the runs the hazard progress method with those values.

Score
Added a GetScore() method to return the satisfaction, dissatisfaction and score variables for each hazard based on an input string. Added a SetFixedBooleans() method to set which booleans are fixed by taking in a string, called by the FinishHazard() method in HazardManager