Streamlined the code a little further
Streamlined the code further so that adding new scripts no longer requires a string variable for it's name. Now the drone raycast script fetches the name of the script that is on the hazard and passes that variable around. Many methods have been simplified to take in less inputs now so are more intuitive to understand and neater looking.

Added a hazard mechanics template script. If added new mechanics it will be much easier to follow this template as it contains all necessary methods/ instructions to implement new hazards.